requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
fsspec==2025.7.0
ftfy==6.3.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
openai-clip==1.0.1
pillow==11.3.0
regex==2025.7.34
requests==2.32.5
setuptools==80.9.0
sympy==1.14.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
